@@ -4,27 +4,26 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
-    "@supabase/supabase-js": "^1.30.5",
+    "@supabase/supabase-js": "^1.33.2",
     "@types/color": "^3.0.3",
-    "axios": "^0.24.0",
+    "axios": "^0.26.1",
     "color": "^4.2.1",
     "command-line-args": "^5.2.1",
-    "dayjs": "^1.10.7",
-    "faunadb": "^4.4.2",
-    "firebase-admin": "^10.0.1",
+    "dayjs": "^1.11.0",
+    "faunadb": "^4.5.4",
     "lodash": "^4.17.21",
-    "rxjs": "^7.5.4",
-    "ts-node": "^10.5.0",
-    "typescript": "^4.5.5",
+    "rxjs": "^7.5.5",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3",
     "winston": "^3.6.0"
   },
   "devDependencies": {
     "@jest/types": "^27.5.1",
-    "@types/jest": "^27.4.0",
-    "@types/lodash": "^4.14.178",
-    "@types/node": "^17.0.19",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^17.0.23",
     "jest": "^27.5.1",
-    "ts-jest": "^27.1.3",
+    "ts-jest": "^27.1.4",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^9.0.0"
